pyproject.toml
[project]
name = "gaugelab"
version = "0.0.0"
authors = [
    { name="Andrew Li", email="andrew@judgmentlabs.ai" },
    { name="Alex Shan", email="alex@judgmentlabs.ai" },
    { name="Joseph Camyre", email="joseph@judgmentlabs.ai" },
]
description = "Gaugelab Package"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
license-files = ["LICENSE.md"]
dependencies = [
    "litellm==1.61.15",
    "python-dotenv==1.0.1",
    "requests",
    "pandas",
    "openai",
    "together",
    "anthropic",
    "nest-asyncio",
    "langchain-huggingface",
    "langchain-openai",
    "langchain-anthropic",
    "langchain-core",
    "google-genai",
    "boto3",
    "matplotlib>=3.10.3",
    "datamodel-code-generator>=0.31.1",
]

[project.urls]
Homepage = "https://github.com/JudgmentLabs/gaugelab"
Issues = "https://github.com/JudgmentLabs/gaugelab/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gaugelab"]
include = [
    "/src/gaugelab",
    "/src/gaugelab/**/*.py",
]

[dependency-groups]
dev = [
    "chromadb>=1.0.12",
    "langchain-community>=0.3.24",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.1",
    "tavily-python>=0.7.5",
    "langgraph>=0.4.3",
    "pre-commit>=4.2.0",
]

[tool.hatch.build]
directory = "dist"
artifacts = [
    "src/gaugelab/**/*.py",
]
exclude = [
    "src/e2etests/*",
    "src/tests/*",
    "src/demo/*"
]

[tool.ruff]
exclude = ["docs"]
